+++ src/curvesimilarities/util.py
@@ -3,8 +3,10 @@
 import functools
 
 import numpy as np
+from scipy.spatial.distance import cdist
 
 __all__ = [
+    "parameter_space",
     "sample_polyline",
     "refine_polyline",
 ]
@@ -39,6 +41,70 @@ def sanitize_vertices(owp):
     return decorator
 
 
+def parameter_space(P, Q, p_num, q_num):
+    r"""Parameter space betwee two polylines.
+
+    Parameters
+    ----------
+    P : array_like
+        A :math:`p` by :math:`n` array of :math:`p` vertices in an
+        :math:`n`-dimensional space.
+    Q : array_like
+        A :math:`q` by :math:`n` array of :math:`q` vertices in an
+        :math:`n`-dimensional space.
+    p_num, q_num : int
+        Number of sample points in `P` and `Q`, respectively.
+
+    Returns
+    -------
+    weight : ndarray
+        A `p_num` by `q_num` array containing the distance between the points in
+        `P` and `Q`.
+    p_coord, q_coord : ndarray
+        Axis coordinates for the parameter space.
+    p_vert, q_vert : ndarray
+        Coordinates for the vertices of polylines.
+
+    Examples
+    --------
+    Curve space:
+
+    .. plot::
+        :context: close-figs
+
+        >>> P = [[0, 0], [2, 2], [4, 2], [4, 4], [2, 1], [5, 1], [7, 2]]
+        >>> Q = [[2, 0], [1, 3], [5, 3], [5, 2], [7, 3]]
+        >>> import matplotlib.pyplot as plt  # doctest: +SKIP
+        >>> plt.plot(*np.array(P).T)  # doctest: +SKIP
+        >>> plt.plot(*np.array(Q).T)  # doctest: +SKIP
+
+    Parameter space with vertices as dashed lines:
+
+    .. plot::
+        :context: close-figs
+
+        >>> weight, p, q, p_vert, q_vert = parameter_space(P, Q, 200, 100)
+        >>> plt.pcolormesh(p, q, weight.T)  # doctest: +SKIP
+        >>> plt.vlines(p_vert, 0, q[-1], "k", "--")  # doctest: +SKIP
+        >>> plt.hlines(q_vert, 0, p[-1], "k", "--")  # doctest: +SKIP
+
+    Free space diagram with Fréchet distance as :math:`\epsilon`:
+
+    .. plot::
+        :context: close-figs
+
+        >>> eps = fd(P, Q)
+        >>> plt.pcolormesh(p, q, weight.T < eps, cmap="gray")  # doctest: +SKIP
+    """
+    p_vert = np.insert(np.cumsum(np.linalg.norm((np.diff(P, axis=0)), axis=-1)), 0, 0)
+    p_coord = np.linspace(0, p_vert[-1], p_num)
+    P_pts = sample_polyline(P, p_coord)
+    q_vert = np.insert(np.cumsum(np.linalg.norm((np.diff(Q, axis=0)), axis=-1)), 0, 0)
+    q_coord = np.linspace(0, q_vert[-1], q_num)
+    Q_pts = sample_polyline(Q, q_coord)
+    return cdist(P_pts, Q_pts), p_coord, q_coord, p_vert, q_vert
+
+
 def sample_polyline(vert, param):
     """Sample points from a polyline.
 
@@ -67,7 +133,7 @@ def sample_polyline(vert, param):
     >>> plt.plot(*np.array(vert).T)  # doctest: +SKIP
     >>> plt.plot(*sample_polyline(vert, param).T, "x")  # doctest: +SKIP
     """
-    vert = np.array(vert)
+    vert = np.asarray(vert)
 
     seg_vec = np.diff(vert, axis=0)
     seg_len = np.linalg.norm(seg_vec, axis=-1)