+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Polyline sampling function `sample_polyline()` in `curvesimilarities.util` module.
 - Polyline refining function `refine_polyline()` in `curvesimilarities.util` module.
+- Parameter space function `parameter_space()` in `curvesimilarities.util` module.
 
 ## [0.1.7] - 2024-06-20